backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
argcomplete==2.0.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==22.12.0
bleach==6.0.0
boltons==23.0.0
cairocffi==1.6.0
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colored==2.2.2
coverage==7.2.7
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.6
docformatter==1.7.3
docutils==0.20.1
eliot==1.14.0
eliot-tree==21.0.0
exceptiongroup==1.1.1
faiss-cpu==1.7.4
fastapi==0.100.0
fastjsonschema==2.17.1
filelock==3.12.2
flake8==6.0.0
frozenlist==1.3.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
griffe==0.30.1
identify==2.5.24
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
iso8601==2.0.0
isort==5.12.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
jupytext==1.14.7
keyring==24.2.0
locusts==0.0.65
lxml==4.9.2
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.1
mkdocs==1.4.3
mkdocs-autorefs==0.4.1
mkdocs-jupyter==0.24.1
mkdocs-material==9.1.17
mkdocs-material-extensions==1.1.1
mkdocstrings==0.22.0
mkdocstrings-python==1.1.2
mknotebooks==0.7.1
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nltk==3.8.1
nodeenv==1.8.0
numpy==1.25.0
openai==0.27.8
packaging==23.1
pandocfilters==1.5.0
pathspec==0.11.1
Pillow==10.0.0
pipx==1.0.0
pkginfo==1.9.6
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.1.2
pyflakes==3.0.1
Pygments==2.15.1
pymdown-extensions==10.0.1
pypdfium2==4.17.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.1.0
readme-renderer==40.0
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
sqlglot==16.8.1
sqlvalidator==0.0.20
starlette==0.27.0
tenacity==8.2.2
textblob==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twine==4.0.2
typer==0.9.0
typing_extensions==4.7.1
untokenize==0.1.1
urllib3==2.0.3
userpath==1.8.0
virtualenv==20.23.1
watchdog==3.0.0
webencodings==0.5.1
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
